Added support for Legacy Launcher Process name ("Infernal Launcher.exe")

+++ LeagueInformerClient/src/main/java/net/nilsghesquiere/runnables/ClientDataManagerRunnable.java
@@ -7,6 +7,7 @@ import java.util.concurrent.TimeUnit;
 
 import net.nilsghesquiere.InfernalBotManagerClient;
 import net.nilsghesquiere.Main;
+import net.nilsghesquiere.util.ProgramConstants;
 import net.nilsghesquiere.util.ProgramUtil;
 
 import org.slf4j.Logger;
@@ -45,19 +46,35 @@ public class ClientDataManagerRunnable implements Runnable {
 					LOGGER.info("No active queuers found, closing InfernalBot process");
 					try {
 						String processName = ProgramUtil.getInfernalProcessname(client.getClientSettings().getInfernalMap());
+						//Check if process is running and kill it (generated name)
 						if(!processName.isEmpty()){
-							ProcessBuilder builder = new ProcessBuilder( "cmd.exe", "/c", "taskkill /F /IM " + processName);
-							builder.redirectErrorStream(true);
-							Process p = builder.start();
-							BufferedReader r = new BufferedReader(new InputStreamReader(p.getInputStream()));
-							String line;
-							while (true) {
-								line = r.readLine();
-								if (line == null) { break; }
-								LOGGER.debug(line);
+							if(ProgramUtil.isProcessRunning(processName)){
+								ProcessBuilder builder = new ProcessBuilder( "cmd.exe", "/c", "taskkill /F /IM " + processName);
+								builder.redirectErrorStream(true);
+								Process p = builder.start();
+								BufferedReader r = new BufferedReader(new InputStreamReader(p.getInputStream()));
+								String line;
+								while (true) {
+									line = r.readLine();
+									if (line == null) { break; }
+									LOGGER.debug(line);
+								}
 							}
 						} else {
 							LOGGER.error("Failure finding current InfernalBot process name");
+						}
+						//Check if legacy process is running and kill it 
+						if(ProgramUtil.isProcessRunning(ProgramConstants.LEGACY_LAUNCHER_NAME)){
+							ProcessBuilder builder2 = new ProcessBuilder( "cmd.exe", "/c", "taskkill /F /IM " + ProgramConstants.LEGACY_LAUNCHER_NAME);
+							builder2.redirectErrorStream(true);
+							Process p2 = builder2.start();
+							BufferedReader r2 = new BufferedReader(new InputStreamReader(p2.getInputStream()));
+							String line2;
+							while (true) {
+								line2 = r2.readLine();
+								if (line2 == null) { break; }
+								LOGGER.debug(line2);
+							}
 						}
 					} catch (IOException e){
 						LOGGER.error("Failure trying to kill InfernalBot Process");
